fix: eslint / prettier style and format rules implementation

+++ app/src/forms/user/service.js
@@ -12,7 +12,7 @@ const service = {
     if (params.idpCode && (params.idpCode === IdentityProviders.BCEIDBASIC || params.idpCode === IdentityProviders.BCEIDBUSINESS)) {
       if (!params.email && !params.username) {
         throw new Problem(422, {
-          detail: 'Could not retrieve BCeID users. Invalid options provided.'
+          detail: 'Could not retrieve BCeID users. Invalid options provided.',
         });
       }
       exact = true;
@@ -31,32 +31,22 @@ const service = {
   },
 
   read: (userId) => {
-    return User.query()
-      .findById(userId)
-      .throwIfNotFound();
+    return User.query().findById(userId).throwIfNotFound();
   },
 
-
   readSafe: (userId) => {
-    return User.query()
-      .modify('safeSelect')
-      .findById(userId)
-      .throwIfNotFound();
+    return User.query().modify('safeSelect').findById(userId).throwIfNotFound();
   },
 
   //
   // User Preferences
   //
   deleteUserPreferences: (currentUser) => {
-    return UserFormPreferences.query()
-      .delete()
-      .where('userId', currentUser.id)
-      .throwIfNotFound();
+    return UserFormPreferences.query().delete().where('userId', currentUser.id).throwIfNotFound();
   },
 
   readUserPreferences: (currentUser) => {
-    return UserFormPreferences.query()
-      .where('userId', currentUser.id);
+    return UserFormPreferences.query().where('userId', currentUser.id);
   },
 
   updateUserPreferences: async (currentUser, body) => {
@@ -64,7 +54,7 @@ const service = {
     try {
       if (!body || !body.forms || !Array.isArray(body.forms)) {
         throw new Problem(422, {
-          detail: 'Could not update user preferences. Invalid options provided'
+          detail: 'Could not update user preferences. Invalid options provided',
         });
       }
 
@@ -74,19 +64,17 @@ const service = {
         const current = await service.readUserFormPreferences(currentUser, form.formId);
 
         if (current) {
-          await UserFormPreferences.query(trx)
-            .patchAndFetchById([currentUser.id, form.formId], {
-              preferences: form.preferences,
-              updatedBy: currentUser.usernameIdp
-            });
+          await UserFormPreferences.query(trx).patchAndFetchById([currentUser.id, form.formId], {
+            preferences: form.preferences,
+            updatedBy: currentUser.usernameIdp,
+          });
         } else {
-          await UserFormPreferences.query(trx)
-            .insert({
-              userId: currentUser.id,
-              formId: form.formId,
-              preferences: form.preferences,
-              createdBy: currentUser.usernameIdp
-            });
+          await UserFormPreferences.query(trx).insert({
+            userId: currentUser.id,
+            formId: form.formId,
+            preferences: form.preferences,
+            createdBy: currentUser.usernameIdp,
+          });
         }
       });
 
@@ -102,15 +90,11 @@ const service = {
   // User Form Preferences
   //
   deleteUserFormPreferences: (currentUser, formId) => {
-    return UserFormPreferences.query()
-      .deleteById([currentUser.id, formId])
-      .throwIfNotFound();
+    return UserFormPreferences.query().deleteById([currentUser.id, formId]).throwIfNotFound();
   },
 
   readUserFormPreferences: (currentUser, formId) => {
-    return UserFormPreferences.query()
-      .findById([currentUser.id, formId])
-      .first();
+    return UserFormPreferences.query().findById([currentUser.id, formId]).first();
   },
 
   updateUserFormPreferences: async (currentUser, formId, preferences) => {
@@ -122,19 +106,17 @@ const service = {
       trx = await UserFormPreferences.startTransaction();
 
       if (current) {
-        result = await UserFormPreferences.query(trx)
-          .patchAndFetchById([currentUser.id, formId], {
-            preferences: preferences,
-            updatedBy: currentUser.usernameIdp
-          });
+        result = await UserFormPreferences.query(trx).patchAndFetchById([currentUser.id, formId], {
+          preferences: preferences,
+          updatedBy: currentUser.usernameIdp,
+        });
       } else {
-        result = await UserFormPreferences.query(trx)
-          .insertAndFetch({
-            userId: currentUser.id,
-            formId: formId,
-            preferences: preferences,
-            createdBy: currentUser.usernameIdp
-          });
+        result = await UserFormPreferences.query(trx).insertAndFetch({
+          userId: currentUser.id,
+          formId: formId,
+          preferences: preferences,
+          createdBy: currentUser.usernameIdp,
+        });
       }
 
       await trx.commit();
@@ -143,7 +125,7 @@ const service = {
       if (trx) await trx.rollback();
       throw err;
     }
-  }
+  },
 };
 
 module.exports = service;
